requirements.txt
gradio>=4.0.0
google-generativeai>=0.3.0
pyyaml>=6.0
pandas>=2.0.0
tenacity>=8.0.0
kagglehub>=0.3.0
numpy>=1.24.0
